Move cache stuff to index::cache
