+++ src/index/cache.rs
@@ -1,4 +1,6 @@
 //! Provides functionality for reading and writing cargo compatible .cache entries
+//! that can be wrapped by another index that has logic for fetching entries
+//! that aren't in the cache
 //!
 //! Cargo creates small cache entries for crates when they are accessed during
 //! any cargo operation that accesses a registry index (update/add/etc).
@@ -9,8 +11,6 @@
 //! they allow skipping network access if in offline mode, as well as allowing
 //! responses from servers to tell the client they have the latest version if
 //! that crate has not been changed since it was last accessed.
-//!
-//!
 //!
 //! +-------------------+---------------------------+------------------+---+
 //! | cache version :u8 | index format version :u32 | revision :string | 0 |
@@ -32,7 +32,7 @@ pub const INDEX_V_MAX: u32 = 2;
 /// The byte representation of [`INDEX_V_MAX`]
 const INDEX_V_MAX_BYTES: [u8; 4] = INDEX_V_MAX.to_le_bytes();
 
-use crate::{CacheError, Error, IndexKrate};
+use crate::{CacheError, Error, IndexKrate, KrateName, PathBuf};
 
 pub struct ValidCacheEntry<'buffer> {
     /// The cache entry's revision
@@ -111,6 +111,7 @@ impl<'buffer> ValidCacheEntry<'buffer> {
 }
 
 impl IndexKrate {
+    /// Reads entries from the versions portion of a cache file
     pub(crate) fn from_cache<'cache>(
         mut iter: impl Iterator<Item = &'cache [u8]> + 'cache,
     ) -> Result<Self, Error> {
@@ -147,8 +148,15 @@ impl IndexKrate {
         w.write_all(revision.as_bytes())?;
         w.write_all(SPLIT)?;
 
+        // crates.io limits crate names to a maximum of 64 characters, but this
+        // only applies to crates.io and not any cargo index, so don't set a hard
+        // limit
+        let mut semver = String::with_capacity(64);
+
         for iv in &self.versions {
-            let semver = iv.version.to_string();
+            semver.clear();
+            // SAFETY: the only way this would fail would be OOM
+            std::fmt::write(&mut semver, format_args!("{}", iv.version)).unwrap();
             w.write_all(semver.as_bytes())?;
             w.write_all(SPLIT)?;
 
@@ -186,4 +194,91 @@ pub fn split(haystack: &[u8], needle: u8) -> impl Iterator<Item = &[u8]> + '_ {
     }
 
     Split { haystack, needle }
+}
+
+/// The [`IndexCache`] allows access to the local cache entries for a remote index
+///
+/// This implementation does no network I/O whatsoever, but does do disk I/O
+pub struct IndexCache {
+    /// The root disk location of the local index
+    pub(super) path: PathBuf,
+}
+
+impl IndexCache {
+    /// Creates a local index exactly at the specified path
+    #[inline]
+    pub fn at_path(path: PathBuf) -> Self {
+        Self { path }
+    }
+
+    /// Reads a crate from the local cache of the index.
+    ///
+    /// You may optionally pass in the revision the cache entry is expected to
+    /// have, if it does match the cache entry will be ignored and an error returned
+    #[inline]
+    pub fn cached_krate(
+        &self,
+        name: KrateName<'_>,
+        revision: Option<&str>,
+    ) -> Result<Option<IndexKrate>, Error> {
+        let Some(contents) = self.read_cache_file(name)? else { return Ok(None) };
+
+        let valid = ValidCacheEntry::read(&contents)?;
+        valid.to_krate(revision)
+    }
+
+    /// Writes the specified crate and revision to the cache
+    pub fn write_to_cache(&self, krate: &IndexKrate, revision: &str) -> Result<PathBuf, Error> {
+        let name = krate.name().try_into()?;
+        let cache_path = self.cache_path(name);
+
+        std::fs::create_dir_all(cache_path.parent().unwrap())?;
+
+        let mut cache_file = match std::fs::File::create(&cache_path) {
+            Ok(cf) => cf,
+            Err(err) => return Err(Error::IoPath(err, cache_path)),
+        };
+
+        // It's unfortunate if this fails for some reason, but
+        // not writing the cache entry shouldn't stop the user
+        // from getting the crate's metadata
+        match krate.write_cache_entry(&mut cache_file, revision) {
+            Ok(_) => Ok(cache_path),
+            Err(err) => {
+                drop(cache_file);
+                // _attempt_ to delete the file, to clean up after ourselves
+                let _ = std::fs::remove_file(&cache_path);
+                Err(Error::IoPath(err, cache_path))
+            }
+        }
+    }
+
+    /// Gets the path the crate's cache file would be located at if it exists
+    #[inline]
+    pub fn cache_path(&self, name: KrateName<'_>) -> PathBuf {
+        let rel_path = name.relative_path(None);
+
+        // avoid realloc on each push
+        let mut cache_path = PathBuf::with_capacity(self.path.as_str().len() + 8 + rel_path.len());
+        cache_path.push(&self.path);
+        cache_path.push(".cache");
+        cache_path.push(rel_path);
+
+        cache_path
+    }
+
+    /// Attempts to read the cache entry for the specified crate
+    ///
+    /// It is recommended to use [`Self::cached_krate`]
+    pub fn read_cache_file(&self, name: KrateName<'_>) -> Result<Option<Vec<u8>>, Error> {
+        let cache_path = self.cache_path(name);
+
+        let cache_bytes = match std::fs::read(&cache_path) {
+            Ok(cb) => cb,
+            Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(None),
+            Err(err) => return Err(Error::IoPath(err, cache_path)),
+        };
+
+        Ok(Some(cache_bytes))
+    }
 }
