+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod cache;
 pub mod error;
 pub mod index;
 pub mod krate;
@@ -7,7 +6,9 @@ pub mod utils;
 pub use camino::{Utf8Path as Path, Utf8PathBuf as PathBuf};
 
 pub use error::{CacheError, Error, HttpError, InvalidUrl, InvalidUrlError};
-pub use index::{git::CRATES_IO_INDEX, sparse::CRATES_IO_HTTP_INDEX};
+pub use index::{
+    git::CRATES_IO_INDEX, sparse::CRATES_IO_HTTP_INDEX, GitIndex, IndexCache, SparseIndex,
+};
 pub use krate::{IndexDependency, IndexKrate, IndexVersion};
 
 #[derive(Copy, Clone)]
